% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{nnd.C.Toep}
\alias{nnd.C.Toep}
\title{nonnegativity check for the given Toeplitz matrix}
\usage{
nnd.C.Toep(Sigma)
}
\arguments{
\item{Sigma}{A matrix to be checked}
}
\value{
It returns the inspection result of the matrix.
}
\description{
nonnegativity check for the given Toeplitz matrix
}
\examples{
nnd.C.Toep(diag(3))
nnd.C.Toep(Sigma.AR.order(0.3, 20))
}
