% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{RBFK}
\alias{RBFK}
\title{RBF kernel function}
\usage{
RBFK(x, y, l)
}
\arguments{
\item{x}{vector input}

\item{y}{vector input (same length with x)}

\item{l}{length parameter in RBF kernel}
}
\value{
asdf
}
\description{
RBF kernel function
}
\examples{
asdf
}
