% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{rmvRBF}
\alias{rmvRBF}
\title{Multivariate normal sampling from the RBF covariance kernel with regular grids}
\usage{
rmvRBF(n, gridpoints, l, mu = rep(0, length(gridpoints)), tau = 1)
}
\arguments{
\item{n}{number of samples extracted from the distribution}

\item{gridpoints}{the gridpoints at which multivariate normal vector is drawn. To ensure the Toeplitz structure of the covariance matrix, the grid should be regular.}

\item{l}{length parameter in the RBF kernel}

\item{mu}{the mean vector of the distribution}

\item{tau}{the constant to be multiplied to all the elements of the covariance matrix. This argument can be used to control the overall variance of the components.}
}
\value{
It returns the list of two. The first element is the n by N (the length of gridpoints) matrix, in which n multivariate normal vectors are stacked vertically.
The second element is the covariance matrix used for the sampling. The argument tau can be used to control the overall variance of the elements.
}
\description{
Multivariate normal sampling from the RBF covariance kernel with regular grids
}
\examples{
rmvRBF(100, c(0:20)/20, l = 0.1, tau = 1)
}
