% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{Sigma.AR.order}
\alias{Sigma.AR.order}
\title{Covariance matrix of an autoregressive model}
\usage{
Sigma.AR.order(rho, p, order = p)
}
\arguments{
\item{rho}{Autocorrelation coefficient; usually the value between 0 and 1}

\item{p}{The number of rows for the desired matrix to be returned}

\item{order}{The order of the AR model; default value is p}
}
\value{
It returns the p by p covariance matrix of an autoregressive (AR) model. By assigning the order, one can customize the number of nonzero elements.
Generated matrices can be used for rmvToep function after checking by nnd.C.Toep function.
The elements in C is calculated as follows.
\eqn{C[i, j] = \rho^{|i-j|}} if \eqn{|i-j|}<=order, else 0
}
\description{
Covariance matrix of an autoregressive model
}
\examples{
Sigma.AR.order(0.3, 10)
Sigma.AR.order(0.5, 10, order = 2)
}
