% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{MatK}
\alias{MatK}
\title{Matern kernel function}
\usage{
MatK(x, y, rho, nu)
}
\arguments{
\item{x}{vector input}

\item{y}{vector input (same length with x)}

\item{rho}{length scale (non-negative)}

\item{nu}{smoothness parameter (non-negative)}
}
\value{
calculated kernel function value
}
\description{
Matern kernel function
}
\examples{
MatK(c(1, 2, 3), c(3, 4, 6), 1, 0.5)
MatK(1, 1, 1, 1)
}
