% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{RBFK}
\alias{RBFK}
\title{RBF kernel function}
\usage{
RBFK(x, y, l)
}
\arguments{
\item{x}{scalar of vector input}

\item{y}{scalar vector input (same length with x)}

\item{l}{length parameter in the RBF kernel}
}
\value{
It returns the calculated RBF kernel function value, along the following formula : \cr
\eqn{C(d) = exp(-\frac{d^2}{2l^2})}, \cr
where \eqn{d = |x-y|}.
}
\description{
RBF kernel function
}
\examples{
RBFK(c(0, 1, 2), 0, 5)
}
