% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{rmvMat}
\alias{rmvMat}
\title{Multivariate normal sampling from RBF covarinace kernel with regular grids}
\usage{
rmvMat(n, gridpoints, rho, nu, mu = rep(0, length(gridpoints)), tau = 1)
}
\arguments{
\item{n}{a}

\item{gridpoints}{b}

\item{rho}{length parameter in Matern kernel}

\item{nu}{smoothness parameter in Matern kernel}

\item{mu}{d}

\item{tau}{e}
}
\value{
asdf
}
\description{
Multivariate normal sampling from RBF covarinace kernel with regular grids
}
\examples{
asdf
}
