% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{rmvMat}
\alias{rmvMat}
\title{Multivariate normal sampling from Matern covarinace kernel with regular grids}
\usage{
rmvMat(n, gridpoints, rho, nu, mu = rep(0, length(gridpoints)), tau = 1)
}
\arguments{
\item{n}{number of samples extracted from the distribution}

\item{gridpoints}{the gridpoints at which multivariate normal vector is drawn. To ensure the Toeplitz structure of the covariance matrix, the grid should be regular.}

\item{rho}{length scale (non-negative)}

\item{nu}{smoothness parameter (non-negative)}

\item{mu}{the mean vector of the distribtuion}

\item{tau}{the constant to be multiplied to all the elements of the covariance matrix}
}
\value{
It returns two values as a list. The first element is the n by N (the length of gridpoints) matrix, in which n multivariate normal vectors are stacked vertically.
The second elemnt is the covariance matrix used for the sampling.
}
\description{
Multivariate normal sampling from Matern covarinace kernel with regular grids
}
\examples{
rmvMat(10, c(0:20)/20, 1, 0.5, tau = 1)
}
