% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{rmvRBF}
\alias{rmvRBF}
\title{Multivariate normal sampling from RBF covarinace kernel with regular grids}
\usage{
rmvRBF(n, gridpoints, l, mu = rep(0, length(gridpoints)), tau = 1)
}
\arguments{
\item{n}{a}

\item{gridpoints}{b}

\item{l}{c}

\item{mu}{d}

\item{tau}{e}
}
\value{
f
}
\description{
Multivariate normal sampling from RBF covarinace kernel with regular grids
}
\examples{
asdf
}
