% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{rmvToep}
\alias{rmvToep}
\title{Title}
\usage{
rmvToep(n, Sigma, mu = NULL, eigtol = 1e-08, symtol = 1e-08)
}
\arguments{
\item{n}{number of samples extracted from the distribution}

\item{Sigma}{covarinace matrix of the distribution. It should be a Toeplitz matrix and the number of rows should be even integer}

\item{mu}{the mean vector of the distribtuion}

\item{eigtol}{numerical threshold for Positive definiteness in Sigma}

\item{symtol}{numerical threshold for checking symmetry in Sigma}
}
\value{
n by N (the number of rows in Sigma) matrix, n multivariate normal vectors stacked vertically
}
\description{
Title
}
\examples{
rmvToep(5, Sigma = diag(10), mu = rep(1, 10))
rmvToep(10, Sigma = diag(abs(norm(6))), mu = rep(1, 6))
}
