% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{rmvToep}
\alias{rmvToep}
\title{Multivariate normal sampling from the Toeplitz-structured Covariance matrix}
\usage{
rmvToep(n, Sigma, mu = rep(0, nrow(Sigma)), tol = 1e-08)
}
\arguments{
\item{n}{number of samples extracted from the distribution}

\item{Sigma}{covariance matrix of the distribution. It should be a Toeplitz matrix}

\item{mu}{the mean vector of the distribution}

\item{tol}{numerical threshold for checking symmetry and Toeplitz structure in Sigma}
}
\value{
It returns the n by N (the number of rows in Sigma) matrix, in which n multivariate normal vectors are stacked vertically. The mean vector of the distribution can be controlled by setting mu vector.
}
\description{
Multivariate normal sampling from the Toeplitz-structured Covariance matrix
}
\examples{
rmvToep(5, Sigma = diag(10), mu = rep(1, 10))
rmvToep(30, Sigma = Sigma.AR.order(0.3, 20))
}
