% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToepMat.R
\name{Sigma.AR.order}
\alias{Sigma.AR.order}
\title{AR structured covariance matrix}
\usage{
Sigma.AR.order(rho, p, order = p)
}
\arguments{
\item{rho}{Autocorrelation coefficient; usually the value between 0 and 1}

\item{p}{The number of rows for the desired matrix to be returned}

\item{order}{The order of the AR model; default value is p}
}
\value{
It returns the p by p matrix with the element. For the return matrix C,
\eqn{C[i, j] = \rho^{|i-j|}} if \eqn{|i-j|<=order}, else 0
}
\description{
AR structured covariance matrix
}
\examples{
Sigma.AR.order(0.3, 20)
Sigma.AR.order(0.5, 10, order = 0)
}
